- Added edit on github link. - Organized Post Components. - Removed React from imports.
