+++ mycomponents/post/index.ts
@@ -1,0 +1,4 @@
+export * from './post-header';
+export * from './post-body';
+export * from './post-cover';
+export * from './post-footer';
