+++ mycomponents/post/post-cover.tsx
@@ -1,0 +1,32 @@
+import Cover from 'types/cover';
+import { ReactElement } from 'react';
+import { Text, Box } from '@chakra-ui/react';
+import NextLink from 'next/link';
+import { ChakraLink } from 'mycomponents/chakra-link';
+import { ChakraNextImage } from 'mycomponents/chakra-next-image';
+
+interface PostCoverProps {
+    cover: Cover;
+}
+function PostCover({ cover }: PostCoverProps): ReactElement {
+    return (
+        <Box mt={8}>
+            <ChakraNextImage
+                src={cover.src}
+                alt={cover.alt}
+                width={1400}
+                height={700}
+                priority
+                objectFit="contain"
+            />
+            <Text align="center">
+                Photo By
+                <NextLink href={cover.author.url} passHref>
+                    <ChakraLink ml={1}>{cover.author.name}</ChakraLink>
+                </NextLink>
+            </Text>
+        </Box>
+    );
+}
+
+export { PostCover };
